requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==44.0.2
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
idna==3.11
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==5.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
